+++ NEWS.md
@@ -5,7 +5,7 @@
 
 * Added a new `summarise_metric` function to the package. This function was previously used internally by the TB report. For a given year it returns the value of given metric, along with the regional and global rankings. The average change over the last decade is also supplied. Linked to #57.
 * `search_data_dict` can now be used to search for a dataset by name. All variables in this dataset are then returned.
-
+* Added a new dataframe, `available_datasets` that lists the datasets available to be imported into R using the package. This dataframe also gives a short description of each dataset, details the timespan of the dataset, and whether or not it is downloaded by default. Used by `get_tb_burden` as a URL source for downloading the datasets. Linked to #58.
 ## Package updates
 
 * `search_data_dict` has improved messaging and no longer returns an error when nothing is found in the data dictionary. From #65.